+++ releases.md
@@ -114,4 +114,5 @@ The module `grants-by-schema` had its granting privileges logic changed. Now we 
 - Module ```grants-by-schema```: Changes the logic of granting privileges to roles
 
 ### Features
-Added resources what apply grants on future objects created in schemas.
+- Added resources what apply grants on future objects created in schemas.
+- Added `snowflake_grant_database_role` resource to grant database roles from schemas to database roles from databases.